--- apps/frontend/src/api/index.ts
@@ -1,1 +1,0 @@
-export { QueryKey } from './constants';
