+++ apps/frontend/src/constants/common.ts
@@ -1,3 +1,2 @@
 export const LOGO_NAME = 'Ease Trip';
-export const AUTHORIZATION_COOKIE_NAME = 'Authorization';
 export const USER_AVATAR_IMAGE_PATH = '/assets/images/male-user-avatar.webp';
