+++ apps/frontend/src/modules/Auth/constants.ts
@@ -1,5 +1,4 @@
 export enum QueryKey {
-  GetUser = 'getUser',
   Login = 'login',
   Register = 'register',
 }
