+++ apps/frontend/src/modules/Home/apis/useGetUser.ts
@@ -1,6 +1,9 @@
 import { useQuery } from '@tanstack/react-query';
-import { QueryKey } from '../constants';
 import { getUser } from './user';
+
+enum QueryKey {
+  GetUser = 'getUser',
+}
 
 export const useGetUser = () => {
   return useQuery({ queryKey: [QueryKey.GetUser], queryFn: getUser });
