+++ apps/frontend/src/modules/Home/components/User.tsx
@@ -1,6 +1,6 @@
 import { useEffect } from 'react';
-import { useGetUser } from '@/api/user';
-import { getUser } from '@/api/user/user';
+import { useGetUser } from '@/modules/Home/apis';
+import { getUser } from '../apis/user';
 
 export function User() {
   const { data } = useGetUser();
